Checkpoint native macOS port and recorded VS validation

+++ src/melee/ft/kinds/ftKirby/ftkirby.c
@@ -2881,7 +2881,7 @@ ftKb_SpecialN_insert_joint_refs(s32* total_dobjs, HSD_Joint* root, Fighter* fp,
             bone++;
             (*part_idx)++;
         }
-        HSD_IDInsertToTable(NULL, (u32) *joint, parts[*part_idx].joint);
+        HSD_IDInsertToTable(NULL, (uintptr_t) *joint, parts[*part_idx].joint);
         (*part_idx)++;
         ftAnim_GetNextJointInTree(joint, joint_idx);
     }
