+++ src/sysdolphin/baselib/id.h
@@ -7,7 +7,7 @@
 
 typedef struct _IDEntry {
     struct _IDEntry* next;
-    u32 id;
+    uintptr_t id;
     void* data;
 } IDEntry;
 
@@ -18,12 +18,12 @@ typedef struct _HSD_IDTable {
 HSD_ObjAllocData* HSD_IDGetAllocData(void);
 void HSD_IDInitAllocData(void);
 void HSD_IDSetup(void);
-void HSD_IDInsertToTable(HSD_IDTable* table, u32 id, void* data);
-void HSD_IDRemoveByIDFromTable(HSD_IDTable* table, u32 id);
-void* HSD_IDGetDataFromTable(HSD_IDTable* table, u32 id, s32* success);
+void HSD_IDInsertToTable(HSD_IDTable* table, uintptr_t id, void* data);
+void HSD_IDRemoveByIDFromTable(HSD_IDTable* table, uintptr_t id);
+void* HSD_IDGetDataFromTable(HSD_IDTable* table, uintptr_t id, s32* success);
 void _HSD_IDForgetMemory(void* low, void* high);
 
-static inline void* HSD_IDGetData(u32 id, s32* success)
+static inline void* HSD_IDGetData(uintptr_t id, s32* success)
 {
     return HSD_IDGetDataFromTable(NULL, id, success);
 }
